+++ client/web/src/integration/graphQlResults.ts
@@ -4,12 +4,6 @@ import { mergeSettings } from '@sourcegraph/shared/src/settings/settings'
 import { testUserID, sharedGraphQlResults } from '@sourcegraph/shared/src/testing/integration/graphQlResults'
 
 import { WebGraphQlOperations } from '../graphql-operations'
-import {
-    collaboratorsPayload,
-    recentFilesPayload,
-    recentSearchesPayload,
-    savedSearchesPayload,
-} from '../search/panels/utils'
 
 import { builtinAuthProvider, siteGQLID, siteID } from './jscontext'
 
@@ -123,8 +117,12 @@ export const commonWebGraphQlResults: Partial<WebGraphQlOperations & SharedGraph
             },
         },
     }),
-    savedSearches: () => ({
-        savedSearches: [],
+    SavedSearches: () => ({
+        savedSearches: {
+            nodes: [],
+            totalCount: 0,
+            pageInfo: { startCursor: null, endCursor: null, hasNextPage: false, hasPreviousPage: false },
+        },
     }),
     LogEvents: () => ({
         logEvents: {
@@ -156,16 +154,6 @@ export const commonWebGraphQlResults: Partial<WebGraphQlOperations & SharedGraph
     }),
     OrgFeatureFlagOverrides: () => ({
         organizationFeatureFlagOverrides: [],
-    }),
-    HomePanelsQuery: () => ({
-        node: {
-            __typename: 'User',
-            recentlySearchedRepositoriesLogs: recentSearchesPayload(),
-            recentSearchesLogs: recentSearchesPayload(),
-            recentFilesLogs: recentFilesPayload(),
-            collaborators: collaboratorsPayload(),
-        },
-        savedSearches: savedSearchesPayload(),
     }),
     SearchHistoryEventLogsQuery: () => ({
         currentUser: {